requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.7.4
django-braces==1.4.0
django-oauth-toolkit==0.8.0
django-postgrespool==0.3.0
djangorestframework==3.1.0
gunicorn==19.2.1
oauthlib==0.7.2
Pillow==2.8.1
psycopg2==2.6
PyJWT==1.1.0
python-social-auth==0.2.6
python3-openid==3.0.5
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
SQLAlchemy==0.9.8
whitenoise==1.0.6
